agronom/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
-e git+https://github.com/musindmitriy/django-debug-panel.git@e7dc30ace1605ee278d6806af13424b93ce481ed#egg=django-debug-panel
-e git+https://github.com/Valera/django-rest-framework-docs.git@support-django-2.0#egg=drfdocs
affine==2.2.2
aioredis==1.2.0
alabaster==0.7.12         # via sphinx
amqp==2.4.1
appnope==0.1.0            # via ipython
async-timeout==3.0.1
atomicwrites==1.3.0       # via pytest
attrs==19.1.0
autopep8==1.4.3           # via django-silk
babel==2.6.0
backcall==0.1.0           # via ipython
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
click-plugins==1.0.4
click==7.0
cligj==0.5.0
cloudpickle==0.8.1
coreapi==2.3.3
coreschema==0.0.4         # via coreapi, drf-yasg
coverage==4.5.2
cycler==0.10.0
dask[array]==1.1.4
decorator==4.4.0
defusedxml==0.5.0
django-allauth==0.39.1
django-cacheops==4.1
django-crispy-forms==1.7.2
django-db-geventpool==3.0.1
django-debug-toolbar==1.11
django-extensions==2.1.5
django-filter==2.1.0
django-leaflet==0.24.0
django-pickling==1.0
django-prometheus==1.0.15
django-silk==3.0.1
django-split-settings==0.3.0
django-webpack-loader==0.6.0
django==2.1.7
djangorestframework-gis==0.14
djangorestframework==3.9.1
docutils==0.14            # via sphinx
drf-extensions==0.4.0
drf-yasg==1.13.0
elasticsearch-dsl==6.3.1
elasticsearch==6.3.1
entrypoints==0.3          # via flake8
flake8==3.7.6
flower==0.9.2
freezegun==0.3.11
funcy==1.11
gevent==1.4.0
gprof2dot==2016.10.13     # via django-silk
greenlet==0.4.15
gunicorn[gevent]==19.9.0
hiredis==1.0.0
httpie==1.0.2
idna==2.8
imagesize==1.1.0          # via sphinx
inflection==0.3.1         # via drf-yasg
ipdb==0.12
ipython-genutils==0.2.0   # via traitlets
ipython==7.4.0            # via ipdb
itypes==1.1.0             # via coreapi
jedi==0.13.3              # via ipython
jinja2==2.10              # via coreschema, django-silk, sphinx
kiwisolver==1.0.1
kombu==4.3.0
markupsafe==1.1.0         # via jinja2
matplotlib==3.0.3
mccabe==0.6.1             # via flake8
more-itertools==6.0.0     # via pytest
networkx==2.2
numpy==1.16.2
oauthlib==3.0.1
packaging==19.0           # via sphinx
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pika==0.13.0
pillow==5.4.1
pip-tools==3.3.2
pluggy==0.8.1             # via pytest
prometheus-client==0.5.0
prompt-toolkit==2.0.9     # via ipython
psycogreen==1.0.1
psycopg2-binary==2.7.7
ptyprocess==0.6.0         # via pexpect
py==1.7.0                 # via pytest
pycodestyle==2.5.0        # via autopep8, flake8
pyflakes==2.1.0           # via flake8
pygments==2.3.1           # via django-silk, httpie, ipython, sphinx
pyparsing==2.3.1
pytest-django==3.4.7
pytest==4.3.0             # via pytest-django
python-dateutil==2.8.0
python-logstash==0.4.6
python3-openid==3.1.0
pytz==2018.9
pywavelets==1.0.2
rasterio==1.0.22
raven==6.10.0
redis==3.2.0
requests-oauthlib==1.2.0
requests==2.21.0
responses==0.10.6
ruamel.yaml==0.15.89      # via drf-yasg
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
snowballstemmer==1.2.1    # via sphinx
snuggs==1.4.3
sphinx-rtd-theme==0.4.3
sphinx==1.8.4
sphinxcontrib-websupport==1.1.0  # via sphinx
sqlparse==0.2.4           # via django-debug-toolbar, django-silk
stories==0.10
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2          # via ipython
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7            # via prompt-toolkit
